+++ example/lib/main.dart
@@ -18,7 +18,7 @@ class MyApp extends StatelessWidget {
         primaryColor: Colors.blue,
         primarySwatch: Colors.blue,
       ),
-      home: Freezed(),
+      home: General(),
       // home: const Music(
       //   cover:
       //       "https://static.wikia.nocookie.net/iowiro/images/c/c5/News_10202019a.jpg",